refactor: name api to utils, adjust exports in index.ts
